+++ CHANGELOG.rst
@@ -4,7 +4,9 @@ Changelog
 0.1.4 - unreleased
 ------------------
 
-
+* Move fallback for verbs, status_codes and paths configuration to ``analyzer``.
+  Also use the fallbacks in :py:meth:`analog.analyzer.Analyzer.__init__` and
+  :py:func:`analog.analyzer.analyze`.
 
 0.1.3 - 2013-01-27
 ------------------